+++ lib/pages/prefs_page.dart
@@ -22,6 +22,7 @@ import 'package:cuppa_mobile/common/padding.dart';
 import 'package:cuppa_mobile/common/platform_adaptive.dart';
 import 'package:cuppa_mobile/common/separators.dart';
 import 'package:cuppa_mobile/common/text_styles.dart';
+import 'package:cuppa_mobile/data/export.dart';
 import 'package:cuppa_mobile/data/localization.dart';
 import 'package:cuppa_mobile/data/prefs.dart';
 import 'package:cuppa_mobile/data/provider.dart';
@@ -145,16 +146,10 @@ class _PrefsWidgetState extends State<PrefsWidget> {
             children: [
               // Setting: teacup style selection
               _cupStyleSetting(context),
-              listDivider,
-              // Setting: collect timer usage stats
-              _collectStatsSetting(context),
-              listDivider,
-              // Setting: use brew ratios
-              _useBrewRatiosSetting(context),
-              listDivider,
+              // Setting: app theme selection
+              _appThemeSetting(context),
               // Setting: show extra info on buttons
               _showExtraSetting(context),
-              listDivider,
               // Setting: stacked timer button view
               Selector<AppProvider, bool>(
                 selector: (_, provider) =>
@@ -162,33 +157,30 @@ class _PrefsWidgetState extends State<PrefsWidget> {
                 builder: (context, showStackedView, child) {
                   return Visibility(
                     visible: showStackedView,
-                    child: Column(
-                      mainAxisSize: MainAxisSize.min,
-                      children: [
-                        _stackedViewSetting(context),
-                        listDivider,
-                      ],
-                    ),
+                    child: _stackedViewSetting(context),
                   );
                 },
               ),
               // Setting: hide timer increment buttons
               _hideIncrementsSetting(context),
               listDivider,
+              // Setting: use brew ratios
+              _useBrewRatiosSetting(context),
+              listDivider,
               // Setting: default to silent timer notifications
               _defaultSilentSetting(context),
+              // Notification info
+              _notificationLink(),
               listDivider,
-              // Setting: app theme selection
-              _appThemeSetting(context),
+              // Setting: collect timer usage stats
+              _collectStatsSetting(context),
+              // Tools: export/import data
+              _exportImportTools(context),
               listDivider,
               // Setting: app language selection
               _appLanguageSetting(context),
-              listDivider,
               // Setting: default to Celsius or Fahrenheit
               _useCelsiusSetting(context),
-              listDivider,
-              // Notification info
-              _notificationLink(),
             ],
           ),
         ),
@@ -428,6 +420,106 @@ class _PrefsWidgetState extends State<PrefsWidget> {
       onChanged: () {
         provider.appLanguage = value;
         Navigator.of(context).pop(true);
+      },
+    );
+  }
+
+  // Tools: export/import data
+  Widget _exportImportTools(BuildContext context) {
+    AppProvider provider = Provider.of<AppProvider>(context);
+
+    return Align(
+      alignment: Alignment.topLeft,
+      child: IgnorePointer(
+        // Disable export/import while timer is active
+        ignoring: provider.activeTeas.isNotEmpty,
+        child: Opacity(
+          opacity: provider.activeTeas.isNotEmpty ? fadeOpacity : noOpacity,
+          child: ListTile(
+            iconColor: Theme.of(context).colorScheme.onPrimaryContainer,
+            title: Text(
+              AppString.export_import.translate(),
+              style: textStyleTitle,
+            ),
+            trailing: Row(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                _exportButton(context),
+                const VerticalDivider(),
+                _importButton(context),
+              ],
+            ),
+            contentPadding: listTilePadding,
+            dense: true,
+          ),
+        ),
+      ),
+    );
+  }
+
+  // Export button
+  Widget _exportButton(BuildContext context) {
+    AppProvider provider = Provider.of<AppProvider>(context);
+
+    return Builder(
+      builder: (BuildContext context) {
+        return IconButton(
+          icon: exportIcon,
+          onPressed: () {
+            // Render location for share sheet on iPad
+            RenderBox? renderBox = context.findRenderObject() as RenderBox?;
+            Rect sharePositionOrigin =
+                (renderBox?.localToGlobal(Offset.zero) ?? Offset.zero) &
+                    (renderBox?.size ?? const Size(4.0, 4.0));
+
+            // Attempt to save an export file and report if failed
+            Export.create(
+              provider,
+              share: true,
+              sharePositionOrigin: sharePositionOrigin,
+            ).then(
+              (exported) {
+                if (!exported && context.mounted) {
+                  showInfoDialog(
+                    context: context,
+                    message: AppString.export_failure.translate(),
+                  );
+                }
+              },
+            );
+          },
+        );
+      },
+    );
+  }
+
+  // Import button
+  Widget _importButton(BuildContext context) {
+    AppProvider provider = Provider.of<AppProvider>(context);
+
+    return IconButton(
+      icon: importIcon,
+      onPressed: () async {
+        // Show a prompt with more information
+        if (await showConfirmDialog(
+          context: context,
+          body: Text(AppString.confirm_import.translate()),
+          bodyExtra: Text(AppString.confirm_continue.translate()),
+        )) {
+          // Attempt to load an export file and report the result
+          Export.load(provider).then(
+            (imported) {
+              if (context.mounted) {
+                showInfoDialog(
+                  context: context,
+                  message: imported
+                      ? AppString.import_sucess.translate()
+                      : AppString.import_failure.translate(),
+                );
+              }
+            },
+          );
+        }
       },
     );
   }
